+++ govee_api_laggat/govee_api_laggat.py
@@ -15,6 +15,7 @@ _API_URL = "https://developer-api.govee.com"
 _RATELIMIT_TOTAL = 'Rate-Limit-Total' # The maximum number of requests you're permitted to make per minute.
 _RATELIMIT_REMAINING = 'Rate-Limit-Remaining' # The number of requests remaining in the current rate limit window.
 _RATELIMIT_RESET = 'Rate-Limit-Reset' # The time at which the current rate limit window resets in UTC epoch seconds.
+BRIGHTNESS_100_MODELS = ["H6163", "H6089","H7022","H6086","H6135","H6137","H7005","H6002","H6003"]
 
 @dataclass
 class GoveeDevice(object):
@@ -39,6 +40,8 @@ class GoveeDeviceState(object):
     power_state: bool
     brightness: int
     color: Tuple[ int, int, int ]
+    timestamp: int
+    source: str
 
 class Govee(object):
     """ client to connect to the govee API """
@@ -57,11 +60,13 @@ class Govee(object):
     def __init__(self, api_key: str):
         """ init with an API_KEY """
         self._api_key = api_key
-        self._devices = []
+        self._devices = {}
+        self._states = {}
         self._rate_limit_on = 5 # safe available call count for multiple processes
         self._limit = 100
         self._limit_remaining = 100
         self._limit_reset = 0
+        self._no_state_before = 0
         
     @classmethod
     async def create(cls, api_key: str):
@@ -75,6 +80,15 @@ class Govee(object):
     def _getAuthHeaders(self):
         return {'Govee-API-Key': self._api_key}
 
+    def _utcnow(self):
+        return datetime.timestamp(datetime.now())
+
+    def _no_state_request_next_seconds(self, sec: int):
+        self._no_state_before = self._utcnow() + sec
+    
+    def _state_request_allowed(self) -> bool:
+        return self._no_state_before < self._utcnow()
+
     def _track_rate_limit(self, response):
         """ rate limiting information """
         if(_RATELIMIT_TOTAL in response.headers and _RATELIMIT_REMAINING in response.headers and _RATELIMIT_RESET in response.headers):
@@ -82,11 +96,15 @@ class Govee(object):
                 self._limit = int(response.headers[_RATELIMIT_TOTAL])
                 self._limit_remaining = int(response.headers[_RATELIMIT_REMAINING])
                 self._limit_reset = float(response.headers[_RATELIMIT_RESET])
+                if response.status == 429:
+                    _LOGGER.warning(f'Rate limit exceeded, check if other devices also utilize the govee API')
             except ex:
                 _LOGGER.warn(f'Cannot track rate limits, response headers: {response.headers}')
 
-    async def _rate_limit(self):
-        if self._limit_remaining <= self._rate_limit_on:
+    async def rate_limit_delay(self):
+        # do we have requests left?
+        if self.rate_limit_remaining <= self.rate_limit_on:
+            # do we need to sleep?
             sleep_sec = self.rate_limit_reset_seconds
             if sleep_sec > 0:
                 _LOGGER.warn(f"Rate limiting active, {self._limit_remaining} of {self._limit} remaining, sleeping for {sleep_sec}s.")
@@ -106,8 +124,7 @@ class Govee(object):
 
     @property
     def rate_limit_reset_seconds(self):
-        utcnow = datetime.timestamp(datetime.now())
-        return self._limit_reset - utcnow
+        return self._limit_reset - self._utcnow()
 
     @property
     def rate_limit_on(self):
@@ -120,12 +137,22 @@ class Govee(object):
         if val < 1:
             raise Exception(f"Rate limiter threshold {val} must be above 1")
         self._rate_limit_on = val
-        
     
     @property
     def devices(self):
         """ returns the cached devices list """
-        return self._devices
+        devs = []
+        for dev in self._devices:
+            devs.append(self._devices[dev])
+        return devs
+    
+    @property
+    def states(self):
+        """ returns the cached states list """
+        devs = []
+        for dev in self._states:
+            devs.append(self._states[dev])
+        return devs
     
     async def ping_async(self) -> Tuple[ float, str ]:
         """ Ping the api endpoint. No API_KEY is needed
@@ -137,7 +164,7 @@ class Govee(object):
         err = None
 
         url = (_API_URL + "/ping")
-        await self._rate_limit()
+        await self.rate_limit_delay()
         async with self._session.get(url=url) as response:
             self._track_rate_limit(response)
             result = await response.text()
@@ -156,20 +183,24 @@ class Govee(object):
     async def get_devices(self) -> Tuple[ List[GoveeDevice], str ]:
         """ get and cache devices, returns: list, error """
         _LOGGER.debug("get_devices")
-        devices = []
+        devices = {}
+        states = {}
         err = None
         
         url = (
             _API_URL
             + "/v1/devices"
         )
-        await self._rate_limit()
+        await self.rate_limit_delay()
         async with self._session.get(url=url, headers = self._getAuthHeaders()) as response:
             self._track_rate_limit(response)
             if response.status == 200:
+                self._no_state_request_next_seconds(2)
                 result = await response.json()
-                devices = [
-                    GoveeDevice(
+                timestamp = self._utcnow()
+                
+                for item in result["data"]["devices"]:
+                    devices[item["device"]] = GoveeDevice(
                         device = item["device"],
                         model = item["model"],
                         device_name = item["deviceName"],
@@ -180,14 +211,24 @@ class Govee(object):
                         support_brightness = "brightness" in item["supportCmds"],
                         support_color = "color" in item["supportCmds"],
                         support_color_tem = "colorTem" in item["supportCmds"]
-                    ) for item in result["data"]["devices"]
-                ]
+                    )
+                    states[item["device"]] = GoveeDeviceState(
+                        device = item["device"],
+                        model = item["model"],
+                        online = True,
+                        power_state = False,
+                        brightness = 0,
+                        color = (0, 0, 0), 
+                        timestamp = timestamp,
+                        source = 'history'
+                    )
             else:
                 result = await response.text()
                 err = f'API-Error {response.status}: {result}'
         # cache last get_devices result
         self._devices = devices
-        return devices, err
+        self._states = states
+        return self.devices, err
 
 
     def _get_device(self, device:  Union[str, GoveeDevice]) -> Tuple[ str, GoveeDevice ]:
@@ -195,10 +236,10 @@ class Govee(object):
         device_str = device
         if isinstance(device, GoveeDevice):
             device_str = device.device
-            if not device in self._devices:
+            if not device_str in self._devices:
                 device = None #disallow unknown devices
-        elif isinstance(device, str):
-            device = next((x for x in self._devices if x.device == device_str), None)
+        elif isinstance(device, str) and device_str in self._devices:
+            device = self._devices[device_str]
         return device_str, device
 
     async def turn_on(self, device: Union[str, GoveeDevice]) -> Tuple[ bool, str ]:
@@ -227,64 +268,77 @@ class Govee(object):
         """
         success = False
         err = None
-        if brightness < 0 or brightness > 254:
-            err = f'set_brightness: invalid value {brightness}, allowed range 0 .. 254'
+        device_str, device = self._get_device(device)
+        if not device:
+            err = f'Invalid device {device_str}, {device}'
         else:
-            brightness_100 = brightness * 100 // 254
-            success, err = await self.set_brightness_100(device, brightness_100)
-        return success, err
-
-    async def set_brightness_100(self, device: Union[str, GoveeDevice], brightness: int) -> Tuple[ bool, str ]:
-        """ set brightness to 0 .. 100 """
-        success = False
-        err = None
-        if brightness < 0 or brightness > 100:
-            err = f'set_brightness100: invalid value {brightness}, allowed range 0 .. 100'
-        else:
-            command = "brightness"
-            result, err = await self._control(device, command, brightness)
-            success = False
-            if not err:
-                success = self._is_success_result_message(result)
+            if brightness < 0 or brightness > 254:
+                err = f'set_brightness: invalid value {brightness}, allowed range 0 .. 254'
+            else:
+                brightness_set = brightness
+                if device.model in BRIGHTNESS_100_MODELS:
+                    brightness_set = brightness * 100 // 254
+                command = "brightness"
+                result, err = await self._control(device, command, brightness_set)
+                if not err:
+                    success = self._is_success_result_message(result)
+                    if success:
+                        self._states[device_str].timestamp = self._utcnow
+                        self._states[device_str].source = 'history'
+                        self._states[device_str].brightness = brightness
         return success, err
 
     async def set_color_temp(self, device: Union[str, GoveeDevice], color_temp: int) -> Tuple[ bool, str ]:
         """ set color temperature to 2000 .. 9000 """
         success = False
         err = None
-        if color_temp < 2000 or color_temp > 9000:
-            err = f'set_color_temp: invalid value {color_temp}, allowed range 2000 .. 9000'
+        device_str, device = self._get_device(device)
+        if not device:
+            err = f'Invalid device {device_str}, {device}'
         else:
-            command = "colorTem"
-            result, err = await self._control(device, command, color_temp)
-            success = False
-            if not err:
-                success = self._is_success_result_message(result)
+            if color_temp < 2000 or color_temp > 9000:
+                err = f'set_color_temp: invalid value {color_temp}, allowed range 2000 .. 9000'
+            else:
+                command = "colorTem"
+                result, err = await self._control(device, command, color_temp)
+                if not err:
+                    success = self._is_success_result_message(result)
+                    if success:
+                        self._states[device_str].timestamp = self._utcnow
+                        self._states[device_str].source = 'history'
+                        self._states[device_str].color_temp = color_temp
         return success, err
 
     async def set_color(self, device: Union[str, GoveeDevice], color: Tuple[ int, int, int ]) -> Tuple[ bool, str ]:
         """ set color (r, g, b) where each value may be in range 0 .. 255 """
         success = False
         err = None
-        if len(color) != 3:
-            err = f'set_color: invalid value {color}, must be tuple with (r, g, b) values'
+        device_str, device = self._get_device(device)
+        if not device:
+            err = f'Invalid device {device_str}, {device}'
         else:
-            red = color[0]
-            green = color[1]
-            blue = color[2]
-            if red < 0 or red > 255:
-                err = f'set_color: invalid value {color}, red must be within 0 .. 254'
-            elif green < 0 or green > 255:
-                err = f'set_color: invalid value {color}, green must be within 0 .. 254'
-            elif blue < 0 or blue > 255:
-                err = f'set_color: invalid value {color}, blue must be within 0 .. 254'
+            if len(color) != 3:
+                err = f'set_color: invalid value {color}, must be tuple with (r, g, b) values'
             else:
-                command = "color"
-                command_color = {"r": red, "g": green, "b": blue}
-                result, err = await self._control(device, command, command_color)
-                success = False
-                if not err:
-                    success = self._is_success_result_message(result)
+                red = color[0]
+                green = color[1]
+                blue = color[2]
+                if red < 0 or red > 255:
+                    err = f'set_color: invalid value {color}, red must be within 0 .. 254'
+                elif green < 0 or green > 255:
+                    err = f'set_color: invalid value {color}, green must be within 0 .. 254'
+                elif blue < 0 or blue > 255:
+                    err = f'set_color: invalid value {color}, blue must be within 0 .. 254'
+                else:
+                    command = "color"
+                    command_color = {"r": red, "g": green, "b": blue}
+                    result, err = await self._control(device, command, command_color)
+                    if not err:
+                        success = self._is_success_result_message(result)
+                        if success:
+                            self._states[device_str].timestamp = self._utcnow
+                            self._states[device_str].source = 'history'
+                            self._states[device_str].color = color
         return success, err
 
     async def _control(self, device: Union[str, GoveeDevice], command: str, params: Any) -> Tuple[ Any, str ]:
@@ -313,7 +367,7 @@ class Govee(object):
                     "model": device.model,
                     "cmd": cmd
                 }
-                await self._rate_limit()
+                await self.rate_limit_delay()
                 async with self._session.put(
                     url=url, 
                     headers = self._getAuthHeaders(),
@@ -321,6 +375,7 @@ class Govee(object):
                 ) as response:
                     self._track_rate_limit(response)
                     if response.status == 200:
+                        self._no_state_request_next_seconds(2)
                         result = await response.json()
                     else:
                         text = await response.text()
@@ -335,8 +390,13 @@ class Govee(object):
         if not device:
             err = f'Invalid device {device_str}'
         elif not device.retrievable:
-            _LOGGER.debug(f"device {device_str} isn't able to return state")
-            pass
+            # device {device_str} isn't able to return state, return 'history' state
+            self._states[device_str].source = 'history'
+            result = self._states[device_str]
+        elif not self._state_request_allowed():
+            # we just changed something, return state from history
+            self._states[device_str].source = 'history'
+            result = self._states[device_str]
         else:
             url = (
                 _API_URL
@@ -346,7 +406,7 @@ class Govee(object):
                 'device': device.device,
                 'model': device.model
             }
-            await self._rate_limit()
+            await self.rate_limit_delay()
             async with self._session.get(
                 url=url,
                 headers = self._getAuthHeaders(),
@@ -354,6 +414,7 @@ class Govee(object):
             ) as response:
                 self._track_rate_limit(response)
                 if response.status == 200:
+                    timestamp = self._utcnow()
                     json_obj = await response.json()
                     prop_online = False
                     prop_power_state = False
@@ -383,8 +444,11 @@ class Govee(object):
                         online = prop_online,
                         power_state = prop_power_state,
                         brightness = prop_brightness,
-                        color = prop_color
+                        color = prop_color, 
+                        timestamp = timestamp,
+                        source = 'api'
                     )
+                    self._states[result.device] = result
                 else:
                     errText = await response.text()
                     err = f'API-Error {response.status}: {result}'