+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+[ ] According to Govee API-support a batch status update is coming - when it is there we could save requests (which are limited to 60/minute)
+  This feature is planned for 2021.
 ### Changed
 ### Deprecated
 ### Removed
@@ -14,7 +16,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.x] - 2020-09-21
 ### Added
-- Support for H6104 (which is controllable, but not retrievable)
+- Support for H6104 and similar devices (which are controllable, but not retrievable)
+- async rate_limit_delay() public available to wait for rate limiting
+### Changed
+- According to Govee API-support brightness is set different on different devices:
+  0-100: "H6089","H7022","H6086","H6135","H6137","H7005","H6002","H6003" (implemented)
+  0-255: all others (this implementation existed)
+  I think this information isn't perfectly correct, my model is 0-100 but not listed. Mail sent.
+- According to Govee API-support we could fix jumping back to old state after controlling device, 
+  by delaying status request for 2-3 seconds. 
+  Status request is answered from cache after any command within two seconds
+### Removed
+- control color by rgb values - not used by home assistant.
 
 ## [0.0.27] - 2020-09-20
 ### Added